Adding my partial attempt at making a launch util
I abandoned this unfinished because I didn't like how it was turning
out. I would have had to launch a python subprocess with the user's
script and arguments manually, which is messy and error-prone, IMO.
Still, I wanted to document this somewhere in case I ever need to refer
back.